+++ src/api/categories.ts
@@ -1,4 +1,6 @@
-export default [
+import { Category } from '../types/Category';
+
+export const categories: Category[] = [
   {
     id: 1,
     title: 'Grocery',
