+++ src/api/users.ts
@@ -1,4 +1,6 @@
-export default [
+import { User } from '../types/User';
+
+export const users: User[] = [
   { id: 1, name: 'Roma', sex: 'm' },
   { id: 2, name: 'Anna', sex: 'f' },
   { id: 3, name: 'Max', sex: 'm' },
