+++ src/types/Category.ts
@@ -1,0 +1,6 @@
+export interface Category {
+  id: number,
+  title: string,
+  icon: string,
+  ownerId: number,
+}
