+++ src/types/ColumnNames.ts
@@ -1,0 +1,6 @@
+export enum ColumnNames {
+  ID = 'ID',
+  Products = 'Product',
+  Category = 'Category',
+  User = 'User',
+}
