+++ tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@mate-academy/students-ts-config",
+  "include": [
+    "src"
+  ]
+}
